feat: Add TypeScript support
Adds the ability to generate TypeScript declarations for PBF-encoded messages.

A new `--typescript` flag has been added to the CLI and as an option to the `compileRaw` function. When enabled, this will produce a TypeScript `.d.ts` file containing interfaces for each of the messages defined in the input protobuf schema.